Separate styles into own file
