+++ src/components/partials/ankle/ankleStyles.js
@@ -1,0 +1,7 @@
+export const root = 'ankle su-relative su-rs-py-7';
+export const heading = 'su-mb-06em su-mt-[-0.6em] su-mx-auto su-max-w-900';
+export const introWrapper =
+  'su-big-paragraph su-max-w-prose su-mx-auto su-text-center su-rs-mb-3';
+export const intro = 'children:su-leading-display';
+export const iconCardGrid =
+  'su-gap-y-2xl md:su-gap-y-[80px] xl:su-gap-y-[100px]';
